Small *.stories refactor
