+++ packages/mr-client/components/models/Labyrinth/Labyrinth.tsx
