+++ packages/mr-client/components/scenes/MainGameScene/MainGameScene.tsx
@@ -2,7 +2,7 @@ import { Loader, OrbitControls } from '@react-three/drei'
 import { Canvas } from '@react-three/fiber'
 import { Suspense } from 'react'
 
-import Labyrinth from '../../models/Labyrinth'
+import Labyrinth from '../../models/Labyrinth/Labyrinth'
 import Lighting from './components/Lighting'
 
 export default function GameScene(): JSX.Element {
